feat: 할인 정책 추가

+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/request/ReserveCreateRequest.java
@@ -11,7 +11,7 @@ public class ReserveCreateRequest {
     private UUID performanceId;
     private String ReservationName; //고객 이름
     private String ReservationPhoneNumber;
-    private long balance;
+    private int balance;
     private int round;
     private int seat;
     private char line;
